requirements.txt
python-telegram-bot>=20.0
pdf2docx>=0.5.8
python-docx>=1.1.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
reportlab>=4.0.0
python-dotenv>=1.0.0
